# Optimizing `gen_image`