--- a/.gitignore
+++ b/.gitignore
@@ -1,1 +1,2 @@
 /docs/build/
+Manifest.toml