Hi @mfiano,

I made a small PR which should increase the performance of `gen_image` a lot.
It's just an optimization around the use of splatting.
This is what I get with `open_simplex2_2d()` (but it's not noise specific) with and without threads

```julia
n = opensimplex2_2d()

# With 4 threads
# main
@btime gen_image($n);
  53.534 ms (5242910 allocations: 184.00 MiB)
# this PR
@btime gen_image($n)
  7.639 ms (27 allocations: 24.00 MiB)

# With 1 thread
# main
@btime gen_image($n);
  168.071 ms (5242890 allocations: 184.00 MiB)
# this PR
 @btime gen_image($n)
  24.538 ms (8 allocations: 24.00 MiB)
```


I also touched the `warp` struct as it was throwing warnings when using the package. The use should be unchanged.